Mocked data autocreation + fixes

+++ orders-service/app/models/product.py
@@ -1,5 +1,6 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String
 from .base import Base
+
 
 class Product(Base):
     __tablename__ = "products"
@@ -8,4 +9,4 @@ class Product(Base):
     name = Column(String, unique=True, index=True, nullable=False)
     description = Column(String, nullable=True)
     stock_quantity = Column(Integer, nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Integer, nullable=False)
